Create PTM_depth_cal.py
hBased on the numerical directory listing possible PTM's per amino acid, the script uses the FASTA file to count the instances of each amino acid in each entry and then uses the numbers to calculate the r. It counts the number of amino acids for each entry (r) by their n (as specified in the directory) and then sum this to derive the PTM-space for each FASTA entry. Then all of these values are summed to derive the proteome-level counts.